Fix various compiler warnings
This is a rather big clean-up patch, after Xlint has been set
to 'all', improving type safety (raw types, unchecked casts),
refactoring some bits where necessary, and fixing some stupid
bugs.

+++ app/src/main/java/ch/unifr/diva/dip/utils/IOUtils.java
@@ -121,7 +121,7 @@ public class IOUtils {
 		if (Files.exists(directory) && isDirectoryEmpty(directory)) {
 			final Path parent = directory.getParent();
 			Files.delete(directory);
-			if (!parent.equals(root)) {
+			if (parent != null && !parent.equals(root)) {
 				deleteDirectoryIfEmpty(parent, root);
 			}
 		}
